refactor(auth): load OAuth marker script on demand
