+++ app-modules/panel-app/src/PanelAppServiceProvider.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace He4rt\PanelApp;
 
+use Filament\Support\Assets\Js;
+use Filament\Support\Facades\FilamentAsset;
 use He4rt\PanelApp\Livewire\Timeline\Composer;
 use He4rt\PanelApp\Livewire\Timeline\Feed;
 use He4rt\PanelApp\Livewire\Timeline\PostShow;
@@ -20,6 +22,11 @@ class PanelAppServiceProvider extends ServiceProvider
     {
         $this->loadViewsFrom(__DIR__.'/../resources/views', 'panel-app');
         $this->loadTranslationsFrom(__DIR__.'/../lang', 'panel-app');
+
+        FilamentAsset::register([
+            Js::make('last-auth-provider', __DIR__.'/../resources/js/last-auth-provider.js')
+                ->loadedOnRequest(),
+        ], package: 'he4rt/panel-app');
 
         Livewire::component('timeline-composer', Composer::class);
         Livewire::component('timeline-feed', Feed::class);
